refactor(canon): split virtual_ts/scope into cohesive modules
