--- crates/vize_canon/src/virtual_ts/scope.rs
@@ -1,1349 +1,0 @@
-//! Scope closure generation for virtual TypeScript.
-//!
-//! Generates TypeScript closures that mirror Vue's template scope hierarchy,
-//! including v-for, v-slot, and event handler scopes. Uses recursive
-//! tree-based generation so nested scopes are properly contained.
-
-use vize_carton::FxHashMap;
-use vize_carton::FxHashSet;
-use vize_carton::String;
-use vize_carton::profile;
-
-use vize_croquis::{
-    BindingMetadata, Croquis, EventHandlerScopeData, Scope, ScopeData, ScopeId, ScopeKind,
-    analysis::ComponentUsage, analyzer::extract_identifiers_oxc, naming::to_pascal_case,
-};
-
-use super::{
-    expressions::{
-        generate_component_prop_checks, generate_expressions, rewrite_reserved_template_prop,
-    },
-    helpers::{
-        generated_text_range, get_dom_event_type, to_camel_case, to_safe_identifier,
-        to_safe_identifier_fragment,
-    },
-    types::{VirtualTsCheckOptions, VirtualTsOptions, VizeMapping},
-};
-use vize_carton::append;
-use vize_carton::cstr;
-
-/// Context for recursive scope generation, bundling shared parameters.
-pub(crate) struct ScopeGenContext<'a> {
-    pub(crate) summary: &'a Croquis,
-    pub(crate) expressions_by_scope: &'a FxHashMap<u32, Vec<&'a vize_croquis::TemplateExpression>>,
-    pub(crate) children_map: &'a FxHashMap<u32, Vec<ScopeId>>,
-    pub(crate) template_prop_names: &'a FxHashSet<String>,
-    pub(crate) template_offset: u32,
-    pub(crate) check_options: VirtualTsCheckOptions,
-}
-
-pub(crate) struct ScopeGenerationOptions<'a> {
-    pub(crate) check_options: VirtualTsCheckOptions,
-    pub(crate) virtual_ts_options: &'a VirtualTsOptions,
-    pub(crate) check_unresolved_global_components: bool,
-}
-
-/// Context for recursive component prop checks inside v-for scopes.
-pub(crate) struct VForPropsContext<'a> {
-    pub(crate) summary: &'a Croquis,
-    pub(crate) components_by_scope: &'a FxHashMap<u32, Vec<(usize, &'a ComponentUsage)>>,
-    pub(crate) children_map: &'a FxHashMap<u32, Vec<ScopeId>>,
-    pub(crate) template_prop_names: &'a FxHashSet<String>,
-    pub(crate) template_offset: u32,
-}
-
-struct EventHandlerExprContext<'a> {
-    expressions_by_scope: &'a FxHashMap<u32, Vec<&'a vize_croquis::TemplateExpression>>,
-    data: &'a EventHandlerScopeData,
-    event_type: &'a str,
-    template_prop_names: &'a FxHashSet<String>,
-    template_offset: u32,
-    indent: &'a str,
-}
-
-struct ComponentPropsContext<'a> {
-    summary: &'a Croquis,
-    children_map: &'a FxHashMap<u32, Vec<ScopeId>>,
-    template_prop_names: &'a FxHashSet<String>,
-    template_offset: u32,
-    options: &'a VirtualTsOptions,
-    check_unresolved_global_components: bool,
-}
-
-/// Generate scope closures from Croquis scope chain.
-/// Uses recursive tree-based generation so nested v-for/v-slot scopes
-/// are properly contained within their parent closures.
-pub(crate) fn generate_scope_closures(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    summary: &Croquis,
-    template_prop_names: &FxHashSet<String>,
-    template_offset: u32,
-    options: ScopeGenerationOptions<'_>,
-) {
-    let check_options = options.check_options;
-    let virtual_ts_options = options.virtual_ts_options;
-
-    // Group expressions by scope_id
-    let expressions_by_scope: FxHashMap<u32, Vec<_>> =
-        profile!("canon.virtual_ts.group_template_expressions", {
-            let mut expressions_by_scope: FxHashMap<u32, Vec<_>> = FxHashMap::default();
-            for expr in &summary.template_expressions {
-                expressions_by_scope
-                    .entry(expr.scope_id.as_u32())
-                    .or_default()
-                    .push(expr);
-            }
-            expressions_by_scope
-        });
-
-    // Build scope tree: parent_scope_id -> Vec<child ScopeId>
-    let children_map: FxHashMap<u32, Vec<ScopeId>> =
-        profile!("canon.virtual_ts.build_scope_tree", {
-            let mut children_map: FxHashMap<u32, Vec<ScopeId>> = FxHashMap::default();
-            for scope in summary.scopes.iter() {
-                if let Some(parent_id) = scope.parent() {
-                    children_map
-                        .entry(parent_id.as_u32())
-                        .or_default()
-                        .push(scope.id);
-                }
-            }
-            children_map
-        });
-
-    // Determine which scopes are nested inside a closure scope (VFor/VSlot).
-    // These will be generated recursively inside their parent, not at top level.
-    let nested_scope_ids: FxHashSet<ScopeId> =
-        profile!("canon.virtual_ts.collect_nested_scope_ids", {
-            summary
-                .scopes
-                .iter()
-                .filter(|scope| {
-                    scope.parent().is_some_and(|pid| {
-                        // Scope ids are arena indices, so resolve the parent
-                        // with the O(1) indexed lookup instead of rescanning
-                        // every scope (was O(n^2) over the scope arena).
-                        summary.scopes.get_scope(pid).is_some_and(|parent| {
-                            matches!(parent.kind, ScopeKind::VFor | ScopeKind::VSlot)
-                        })
-                    })
-                })
-                .map(|scope| scope.id)
-                .collect()
-        });
-
-    if check_options.check_template_bindings {
-        profile!(
-            "canon.virtual_ts.instance_global_refs",
-            generate_instance_global_refs(
-                ts,
-                mappings,
-                summary,
-                template_offset,
-                virtual_ts_options
-            )
-        );
-    }
-
-    // Process non-nested scopes at template level
-    for scope in summary.scopes.iter() {
-        let scope_id = scope.id.as_u32();
-
-        // Skip scopes that are nested inside a closure parent
-        if nested_scope_ids.contains(&scope.id) {
-            continue;
-        }
-
-        // Global scopes: emit expressions directly
-        if matches!(
-            scope.kind,
-            ScopeKind::JsGlobalUniversal
-                | ScopeKind::JsGlobalBrowser
-                | ScopeKind::JsGlobalNode
-                | ScopeKind::VueGlobal
-        ) {
-            if let Some(exprs) = expressions_by_scope.get(&scope_id)
-                && check_options.check_template_bindings
-            {
-                generate_expressions(
-                    ts,
-                    mappings,
-                    exprs,
-                    template_prop_names,
-                    template_offset,
-                    "  ",
-                );
-            }
-            continue;
-        }
-
-        let ctx = ScopeGenContext {
-            summary,
-            expressions_by_scope: &expressions_by_scope,
-            children_map: &children_map,
-            template_prop_names,
-            template_offset,
-            check_options,
-        };
-        profile!(
-            "canon.virtual_ts.scope_node",
-            generate_scope_node(ts, mappings, &ctx, scope, "  ")
-        );
-    }
-
-    // Handle undefined references
-    if check_options.check_template_bindings {
-        profile!(
-            "canon.virtual_ts.undefined_refs",
-            generate_undefined_refs(ts, mappings, summary, template_offset)
-        );
-    }
-
-    // Generate component props type checks (scope-aware)
-    if check_options.check_props {
-        profile!(
-            "canon.virtual_ts.component_props",
-            generate_component_props(
-                ts,
-                mappings,
-                &ComponentPropsContext {
-                    summary,
-                    children_map: &children_map,
-                    template_prop_names,
-                    template_offset,
-                    options: virtual_ts_options,
-                    check_unresolved_global_components: options.check_unresolved_global_components,
-                },
-            )
-        );
-    }
-}
-
-/// Handle undefined references from template.
-fn generate_undefined_refs(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    summary: &Croquis,
-    template_offset: u32,
-) {
-    if summary.undefined_refs.is_empty() {
-        return;
-    }
-
-    // Collect type export names to exclude from undefined refs
-    let type_export_names: FxHashSet<&str> = summary
-        .type_exports
-        .iter()
-        .map(|te| te.name.as_str())
-        .collect();
-
-    let mut seen_names: FxHashSet<&str> = FxHashSet::default();
-    let mut emitted_header = false;
-    for undef in &summary.undefined_refs {
-        if !seen_names.insert(undef.name.as_str()) {
-            continue;
-        }
-        if is_template_instance_global_name(undef.name.as_str()) {
-            continue;
-        }
-        // Skip names that match type exports (these are type-level, not value-level)
-        if type_export_names.contains(undef.name.as_str()) {
-            continue;
-        }
-
-        let src_start = (template_offset + undef.offset) as usize;
-        let src_end = src_start + undef.name.len();
-
-        if !emitted_header {
-            ts.push_str("\n  // Undefined references from template:\n");
-            emitted_header = true;
-        }
-
-        let gen_start = ts.len();
-        // Use void expression to reference the name without creating an unused variable
-        let expr_code = cstr!("  void ({});\n", undef.name);
-        let name_offset = expr_code.find(undef.name.as_str()).unwrap_or(0);
-        let gen_name_start = gen_start + name_offset;
-        let gen_name_end = gen_name_start + undef.name.len();
-
-        ts.push_str(&expr_code);
-        mappings.push(VizeMapping {
-            gen_range: gen_name_start..gen_name_end,
-            src_range: src_start..src_end,
-            sub_spans: Vec::new(),
-        });
-        append!(
-            *ts,
-            "  // @vize-map: {gen_name_start}:{gen_name_end} -> {src_start}:{src_end}\n",
-        );
-    }
-}
-
-fn generate_instance_global_refs(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    summary: &Croquis,
-    template_offset: u32,
-    options: &VirtualTsOptions,
-) {
-    if summary.undefined_refs.is_empty() && summary.template_expressions.is_empty() {
-        return;
-    }
-
-    let mut emitter = InstanceGlobalRefsEmitter::new(ts, mappings, summary, options);
-    for undef in &summary.undefined_refs {
-        let src_start = (template_offset + undef.offset) as usize;
-        let src_end = src_start + undef.name.len();
-        emitter.emit(undef.name.as_str(), src_start, src_end);
-    }
-
-    for expr in &summary.template_expressions {
-        for ident in extract_identifiers_oxc(expr.content.as_str()) {
-            let name = ident.as_str();
-            let Some(relative_offset) = expr.content.find(name) else {
-                continue;
-            };
-            let src_start = (template_offset + expr.start) as usize + relative_offset;
-            let src_end = src_start + name.len();
-            emitter.emit(name, src_start, src_end);
-        }
-    }
-}
-
-struct InstanceGlobalRefsEmitter<'a> {
-    ts: &'a mut String,
-    mappings: &'a mut Vec<VizeMapping>,
-    options: &'a VirtualTsOptions,
-    bindings: &'a BindingMetadata,
-    type_export_names: FxHashSet<&'a str>,
-    seen_names: FxHashSet<String>,
-    emitted_header: bool,
-}
-
-impl<'a> InstanceGlobalRefsEmitter<'a> {
-    fn new(
-        ts: &'a mut String,
-        mappings: &'a mut Vec<VizeMapping>,
-        summary: &'a Croquis,
-        options: &'a VirtualTsOptions,
-    ) -> Self {
-        Self {
-            ts,
-            mappings,
-            options,
-            bindings: &summary.bindings,
-            type_export_names: summary
-                .type_exports
-                .iter()
-                .map(|te| te.name.as_str())
-                .collect(),
-            seen_names: FxHashSet::default(),
-            emitted_header: false,
-        }
-    }
-
-    fn emit(&mut self, name: &str, src_start: usize, src_end: usize) {
-        if !is_template_instance_global_name(name)
-            || self.bindings.contains(name)
-            || self.type_export_names.contains(name)
-            || is_declared_template_context_name(name, self.options)
-            || !self.seen_names.insert(name.into())
-        {
-            return;
-        }
-
-        if !self.emitted_header {
-            self.ts
-                .push_str("\n  // Instance globals from ComponentPublicInstance:\n");
-            self.ts.push_str(
-                "  type __VizeInstanceGlobal<K extends string> = K extends keyof __Ctx ? __Ctx[K] : any;\n",
-            );
-            self.emitted_header = true;
-        }
-
-        let gen_start = self.ts.len();
-        let stmt = cstr!("  const {name}: __VizeInstanceGlobal<'{name}'> = undefined as any;\n");
-        let gen_name_start = gen_start + stmt.find(name).unwrap_or(0);
-        let gen_name_end = gen_name_start + name.len();
-
-        self.ts.push_str(&stmt);
-        self.mappings.push(VizeMapping {
-            gen_range: gen_name_start..gen_name_end,
-            src_range: src_start..src_end,
-            sub_spans: Vec::new(),
-        });
-        append!(
-            *self.ts,
-            "  // @vize-map: {gen_name_start}:{gen_name_end} -> {src_start}:{src_end}\n",
-        );
-    }
-}
-
-fn is_template_instance_global_name(name: &str) -> bool {
-    let Some(rest) = name.strip_prefix('$') else {
-        return false;
-    };
-    !rest.is_empty()
-        && rest
-            .chars()
-            .all(|c| c == '_' || c == '$' || c.is_ascii_alphanumeric())
-}
-
-fn is_declared_template_context_name(name: &str, options: &VirtualTsOptions) -> bool {
-    matches!(name, "$attrs" | "$slots" | "$refs" | "$emit" | "$event")
-        || options
-            .template_globals
-            .iter()
-            .any(|global| global.name.as_str() == name)
-        || options
-            .css_modules
-            .iter()
-            .any(|module_name| module_name.as_str() == name)
-}
-
-/// Type annotation for a `v-slot` scope's props. When the slot is on a child
-/// component (`component` is `Some`), the props are inferred from that child's
-/// `$slots[name]` parameter (its `defineSlots`), so misuse raises a real
-/// diagnostic (#764). Dynamic slot names are matched against the union of all
-/// declared slot function props, matching Vue's runtime lookup without
-/// treating the expression text as a static slot key. Otherwise — and whenever
-/// the child has no typed slot — it falls back to `any` so untyped or built-in
-/// slot hosts never produce a false positive.
-fn slot_props_type(component: Option<&str>, slot_name: &str, slot_name_is_static: bool) -> String {
-    match component {
-        Some(component) => {
-            let component_ref = to_safe_identifier(component);
-            if slot_name_is_static {
-                cstr!(
-                    "typeof {component_ref} extends {{ new (): {{ $slots: infer __S }} }} ? (__S extends {{ \"{slot_name}\"?: (props: infer __P, ...args: any[]) => any }} ? __P : any) : any"
-                )
-            } else {
-                cstr!(
-                    "typeof {component_ref} extends {{ new (): {{ $slots: infer __S }} }} ? ({{ [__K in keyof __S]: NonNullable<__S[__K]> extends (props: infer __P, ...args: any[]) => any ? __P : never }}[keyof __S] extends infer __P ? ([__P] extends [never] ? any : __P) : any) : any"
-                )
-            }
-        }
-        None => "any".into(),
-    }
-}
-
-fn append_v_for_comment(ts: &mut String, indent: &str, label: &str, alias: &str, source: &str) {
-    append!(*ts, "\n{indent}// {label}: {alias} in ");
-    for c in source.chars() {
-        if c == '\n' || c == '\r' {
-            ts.push(' ');
-        } else {
-            ts.push(c);
-        }
-    }
-    ts.push('\n');
-}
-
-/// Emit the opening of a v-for scope as
-/// `__vForList(source).forEach(([value, key, index]) => {`.
-///
-/// The overloaded `__vForList` helper types the destructured tuple from the
-/// source kind: arrays/iterables/numbers/strings keep a numeric `key`, while an
-/// object source yields `value: T[keyof T]` and `key: keyof T` (matching
-/// vue-tsc) instead of the old array-only `(source).forEach` assumption that
-/// mis-typed objects and raised spurious TS2339/TS2537. The source expression is
-/// passed through verbatim so any `as Type` assertion flows into the helper.
-fn emit_v_for_loop_open(
-    ts: &mut String,
-    indent: &str,
-    value_alias: &str,
-    key_alias: Option<&str>,
-    index_alias: Option<&str>,
-    source: &str,
-) {
-    append!(*ts, "{indent}__vForList({source}).forEach(([{value_alias}");
-    if let Some(key) = key_alias {
-        append!(*ts, ", {key}");
-    } else if index_alias.is_some() {
-        // Keep the index in the third tuple slot even when no key alias is bound.
-        ts.push_str(", _key");
-    }
-    if let Some(index) = index_alias {
-        append!(*ts, ", {index}");
-    }
-    ts.push_str("]) => {\n");
-}
-
-/// Generate component props type checks (scope-aware).
-/// Type declarations are at template level, value checks are in their scope.
-fn generate_component_props(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    ctx: &ComponentPropsContext<'_>,
-) {
-    let summary = ctx.summary;
-    if summary.component_usages.is_empty() {
-        return;
-    }
-
-    let external_template_bindings: FxHashSet<&str> = ctx
-        .options
-        .external_template_bindings
-        .iter()
-        .map(|name| name.as_str())
-        .collect();
-    let checkable_usages: Vec<(usize, &ComponentUsage)> = summary
-        .component_usages
-        .iter()
-        .enumerate()
-        .filter(|(_, usage)| {
-            component_usage_has_checkable_binding(
-                summary,
-                usage,
-                &external_template_bindings,
-                ctx.check_unresolved_global_components,
-            )
-        })
-        .collect();
-    if checkable_usages.is_empty() {
-        return;
-    }
-
-    // Group component usages by scope_id
-    let mut components_by_scope: FxHashMap<u32, Vec<(usize, &ComponentUsage)>> =
-        FxHashMap::default();
-    for &(idx, usage) in &checkable_usages {
-        components_by_scope
-            .entry(usage.scope_id.as_u32())
-            .or_default()
-            .push((idx, usage));
-    }
-
-    // Emit type declarations only for components with dynamic props
-    // (TypeScript type aliases cannot be inside function bodies)
-    ts.push_str("\n  // Component props type declarations\n");
-
-    // Helper types for the generic functional prop-check path (#775). A
-    // `<script setup generic="T">` child exposes `__vizeCheck<T>(props)` on its
-    // default export; `__VizePropChecker<C>` extracts that generic signature so
-    // the parent can call it and let TS infer `T` from the passed props. When
-    // the child is non-generic (plain construct signature), a built-in / library
-    // component, or `any`, it falls back to `(props: any) => void`, a no-op that
-    // never reports, so only generic components take the new path and the
-    // well-tested `typeof Comp extends { $props }` extraction below is preserved.
-    let any_dynamic_props = checkable_usages.iter().any(|(_, usage)| {
-        usage.props.iter().any(|p| {
-            p.name.as_str() != "key"
-                && p.name.as_str() != "ref"
-                && p.value.is_some()
-                && p.is_dynamic
-        })
-    });
-    if any_dynamic_props {
-        ts.push_str("  type __VizeIsAny<T> = 0 extends (1 & T) ? true : false;\n");
-        ts.push_str(
-            "  type __VizePropChecker<C> = __VizeIsAny<C> extends true ? (props: any) => void : C extends { __vizeCheck: infer __F } ? (__F extends (...args: any[]) => any ? __F : (props: any) => void) : (props: any) => void;\n",
-        );
-        ts.push_str(
-            "  type __VizePropValue<P, K extends PropertyKey> = K extends keyof P ? P[K] : unknown;\n",
-        );
-    }
-
-    for &(idx, usage) in &checkable_usages {
-        let component_ref = to_safe_identifier(usage.name.as_str());
-        let component_type_name = to_safe_identifier_fragment(usage.name.as_str());
-
-        // Only emit type when there are dynamic props to check
-        let has_dynamic_props = usage.props.iter().any(|p| {
-            p.name.as_str() != "key"
-                && p.name.as_str() != "ref"
-                && p.value.is_some()
-                && p.is_dynamic
-        });
-        if !has_dynamic_props {
-            continue;
-        }
-
-        let src_start = (ctx.template_offset + usage.start) as usize;
-        let src_end = (ctx.template_offset + usage.end) as usize;
-
-        append!(*ts, "  // @vize-map: component -> {src_start}:{src_end}\n",);
-        append!(
-            *ts,
-            "  type __{component_type_name}_Props_{idx} = typeof {component_ref} extends {{ new (): {{ $props: infer __P }} }} ? __P : (typeof {component_ref} extends (props: infer __P) => any ? __P : {{}});\n",
-        );
-
-        for prop in &usage.props {
-            if prop.name.as_str() == "key" || prop.name.as_str() == "ref" {
-                continue;
-            }
-            if prop.value.is_some() && prop.is_dynamic {
-                let camel_prop_name = to_camel_case(prop.name.as_str());
-                let safe_prop_name = to_safe_identifier_fragment(prop.name.as_str());
-                append!(
-                    *ts,
-                    "  type __{component_type_name}_{idx}_prop_{safe_prop_name} = __VizePropValue<__{component_type_name}_Props_{idx}, '{camel_prop_name}'>;\n",
-                );
-            }
-        }
-
-        // Generic functional prop-checker for this component (#775). Resolves to
-        // the child's `__vizeCheck` when generic, else a `(props: any)` no-op.
-        append!(
-            *ts,
-            "  type __{component_type_name}_Check_{idx} = __VizePropChecker<typeof {component_ref}>;\n",
-        );
-    }
-
-    // Collect all closure scope IDs (v-for and v-slot)
-    let closure_scope_ids: FxHashSet<u32> = summary
-        .scopes
-        .iter()
-        .filter(|s| matches!(s.kind, ScopeKind::VFor | ScopeKind::VSlot))
-        .map(|s| s.id.as_u32())
-        .collect();
-
-    // Root closure scopes: VFor/VSlot scopes whose parent is NOT a closure scope
-    let root_closure_scope_ids: FxHashSet<u32> = summary
-        .scopes
-        .iter()
-        .filter(|s| {
-            matches!(s.kind, ScopeKind::VFor | ScopeKind::VSlot)
-                && s.parent().is_none_or(|pid| {
-                    // O(1) arena lookup of the parent scope rather than a
-                    // linear find per scope (was O(n^2) over the arena).
-                    summary
-                        .scopes
-                        .get_scope(pid)
-                        .is_none_or(|p| !matches!(p.kind, ScopeKind::VFor | ScopeKind::VSlot))
-                })
-        })
-        .map(|s| s.id.as_u32())
-        .collect();
-
-    ts.push_str("\n  // Component props value checks (template scope)\n");
-    for &(idx, usage) in &checkable_usages {
-        if closure_scope_ids.contains(&usage.scope_id.as_u32()) {
-            continue; // Will be emitted inside v-for/v-slot scope
-        }
-        profile!(
-            "canon.virtual_ts.component_prop_checks",
-            generate_component_prop_checks(
-                ts,
-                mappings,
-                usage,
-                idx,
-                ctx.template_prop_names,
-                ctx.template_offset,
-                "  "
-            )
-        );
-    }
-
-    // Emit value checks for components in closure scopes (v-for and v-slot)
-    for scope in summary.scopes.iter() {
-        if !matches!(scope.kind, ScopeKind::VFor | ScopeKind::VSlot) {
-            continue;
-        }
-        // Only process root closure scopes; nested ones are handled recursively
-        if !root_closure_scope_ids.contains(&scope.id.as_u32()) {
-            continue;
-        }
-        let props_ctx = VForPropsContext {
-            summary,
-            components_by_scope: &components_by_scope,
-            children_map: ctx.children_map,
-            template_prop_names: ctx.template_prop_names,
-            template_offset: ctx.template_offset,
-        };
-        profile!(
-            "canon.virtual_ts.closure_component_props",
-            generate_closure_component_props_recursive(ts, mappings, &props_ctx, scope, "  ")
-        );
-    }
-}
-
-fn component_usage_has_checkable_binding(
-    summary: &Croquis,
-    usage: &ComponentUsage,
-    external_template_bindings: &FxHashSet<&str>,
-    check_unresolved_global_components: bool,
-) -> bool {
-    let name = usage.name.as_str();
-    summary.bindings.bindings.contains_key(name)
-        || external_template_bindings.contains(name)
-        || (check_unresolved_global_components && !name.is_empty())
-}
-
-/// Recursively generate a scope node (VFor/VSlot/EventHandler) and its nested children.
-fn generate_scope_node(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    ctx: &ScopeGenContext<'_>,
-    scope: &Scope,
-    indent: &str,
-) {
-    let scope_id = scope.id.as_u32();
-    let inner_indent = cstr!("{indent}  ");
-
-    match scope.data() {
-        ScopeData::VFor(data) => {
-            append_v_for_comment(
-                ts,
-                indent,
-                "v-for scope",
-                data.value_alias.as_str(),
-                data.source.as_str(),
-            );
-
-            emit_v_for_loop_open(
-                ts,
-                indent,
-                data.value_alias.as_str(),
-                data.key_alias.as_deref(),
-                data.index_alias.as_deref(),
-                data.source.as_str(),
-            );
-
-            // Mark v-for variables as used to avoid TS6133
-            for value in &data.value_bindings {
-                append!(*ts, "{inner_indent}void {value};\n");
-            }
-            if let Some(ref key) = data.key_alias {
-                append!(*ts, "{inner_indent}void {key};\n");
-            }
-            if let Some(ref index) = data.index_alias {
-                append!(*ts, "{inner_indent}void {index};\n");
-            }
-
-            // Generate expressions in this scope
-            if let Some(exprs) = ctx.expressions_by_scope.get(&scope_id)
-                && ctx.check_options.check_template_bindings
-            {
-                generate_expressions(
-                    ts,
-                    mappings,
-                    exprs,
-                    ctx.template_prop_names,
-                    ctx.template_offset,
-                    &inner_indent,
-                );
-            }
-
-            // Recursively generate child scopes inside this closure
-            profile!(
-                "canon.virtual_ts.child_scopes",
-                generate_child_scopes(ts, mappings, ctx, scope_id, &inner_indent)
-            );
-
-            ts.push_str(indent);
-            ts.push_str("});\n");
-        }
-        ScopeData::VSlot(data) => {
-            append!(*ts, "\n{indent}// v-slot scope: #{}\n", data.name);
-
-            let props_pattern = data.props_pattern.as_deref().unwrap_or("slotProps");
-            let safe_slot_name = to_safe_identifier_fragment(data.name.as_str());
-            let props_type = slot_props_type(
-                data.component.as_deref(),
-                data.name.as_str(),
-                ctx.summary.scopes.is_v_slot_name_static(scope.id),
-            );
-            append!(
-                *ts,
-                "{indent}void function _slot_{safe_slot_name}({props_pattern}: {props_type}) {{\n",
-            );
-            // Mark slot prop variables as used
-            if data.prop_names.is_empty() {
-                // Simple identifier (no destructuring)
-                append!(*ts, "{inner_indent}void {props_pattern};\n");
-            } else {
-                // Destructured: void each extracted prop name
-                for prop_name in data.prop_names.iter() {
-                    append!(*ts, "{inner_indent}void {prop_name};\n");
-                }
-            }
-
-            if let Some(exprs) = ctx.expressions_by_scope.get(&scope_id)
-                && ctx.check_options.check_template_bindings
-            {
-                generate_expressions(
-                    ts,
-                    mappings,
-                    exprs,
-                    ctx.template_prop_names,
-                    ctx.template_offset,
-                    &inner_indent,
-                );
-            }
-
-            // Recursively generate child scopes inside this closure
-            profile!(
-                "canon.virtual_ts.child_scopes",
-                generate_child_scopes(ts, mappings, ctx, scope_id, &inner_indent)
-            );
-
-            ts.push_str(indent);
-            ts.push_str("};\n");
-        }
-        ScopeData::EventHandler(data) if ctx.check_options.check_emits => {
-            append!(*ts, "\n{indent}// @{} handler\n", data.event_name);
-
-            let safe_event_name = to_safe_identifier(data.event_name.as_str());
-
-            if let Some(ref component_name) = data.target_component {
-                let component_ref = to_safe_identifier(component_name.as_str());
-                let component_type_name = to_safe_identifier_fragment(component_name.as_str());
-                let pascal_event = to_pascal_case(data.event_name.as_str());
-                let on_handler = cstr!("on{pascal_event}");
-
-                let prop_key = if on_handler.contains(':') {
-                    cstr!("\"{on_handler}\"")
-                } else {
-                    on_handler
-                };
-
-                // Type alias (block-scoped in TypeScript)
-                // Include scope_id to deduplicate when same component+event appears multiple times
-                append!(
-                    *ts,
-                    "{indent}type __{component_type_name}_{scope_id}_{safe_event_name}_args = typeof {component_ref} extends {{ new (): {{ $props: infer __P }} }}\n",
-                );
-                append!(
-                    *ts,
-                    "{indent}  ? __P extends {{ {prop_key}?: (...args: infer __A) => any }} ? __A : unknown[]\n",
-                );
-                append!(
-                    *ts,
-                    "{indent}  : typeof {component_ref} extends (props: infer __P) => any\n",
-                );
-                append!(
-                    *ts,
-                    "{indent}    ? __P extends {{ {prop_key}?: (...args: infer __A) => any }} ? __A : unknown[]\n",
-                );
-                append!(*ts, "{indent}    : unknown[];\n");
-                append!(
-                    *ts,
-                    "{indent}type __{component_type_name}_{scope_id}_{safe_event_name}_event = __{component_type_name}_{scope_id}_{safe_event_name}_args extends [] ? any : __{component_type_name}_{scope_id}_{safe_event_name}_args[0];\n",
-                );
-
-                let event_type =
-                    cstr!("__{component_type_name}_{scope_id}_{safe_event_name}_event");
-                append!(*ts, "{indent}(($event: {event_type}) => {{\n");
-
-                profile!(
-                    "canon.virtual_ts.event_handler_expressions",
-                    generate_event_handler_expressions(
-                        ts,
-                        mappings,
-                        scope_id,
-                        &EventHandlerExprContext {
-                            expressions_by_scope: ctx.expressions_by_scope,
-                            data,
-                            event_type: event_type.as_str(),
-                            template_prop_names: ctx.template_prop_names,
-                            template_offset: ctx.template_offset,
-                            indent: &inner_indent,
-                        },
-                    )
-                );
-
-                append!(*ts, "{indent}}})({{}} as {event_type});\n");
-            } else {
-                let event_type = get_dom_event_type(data.event_name.as_str());
-                append!(*ts, "{indent}(($event: {event_type}) => {{\n");
-
-                profile!(
-                    "canon.virtual_ts.event_handler_expressions",
-                    generate_event_handler_expressions(
-                        ts,
-                        mappings,
-                        scope_id,
-                        &EventHandlerExprContext {
-                            expressions_by_scope: ctx.expressions_by_scope,
-                            data,
-                            event_type,
-                            template_prop_names: ctx.template_prop_names,
-                            template_offset: ctx.template_offset,
-                            indent: &inner_indent,
-                        },
-                    )
-                );
-
-                append!(*ts, "{indent}}})({{}} as {event_type});\n");
-            }
-        }
-        _ => {
-            if let Some(exprs) = ctx.expressions_by_scope.get(&scope_id)
-                && ctx.check_options.check_template_bindings
-            {
-                generate_expressions(
-                    ts,
-                    mappings,
-                    exprs,
-                    ctx.template_prop_names,
-                    ctx.template_offset,
-                    indent,
-                );
-            }
-        }
-    }
-}
-
-/// Generate event handler expressions inside a closure.
-fn generate_event_handler_expressions(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    scope_id: u32,
-    ctx: &EventHandlerExprContext<'_>,
-) {
-    if let Some(exprs) = ctx.expressions_by_scope.get(&scope_id) {
-        for expr in exprs {
-            let content = expr.content.as_str();
-            let is_implicit_reference =
-                ctx.data.has_implicit_event && is_callable_handler_reference(content);
-            let inline_callback_event_arg = inline_callback_event_argument(content);
-            let src_start = (ctx.template_offset + expr.start) as usize;
-            let src_end = (ctx.template_offset + expr.end) as usize;
-            let guard = expr.vif_guard.as_ref().map(|guard| {
-                let trimmed_guard = guard.as_str().trim();
-                rewrite_reserved_template_prop(trimmed_guard, ctx.template_prop_names)
-                    .unwrap_or_else(|| String::from(guard.as_str()))
-            });
-            if let Some(ref guard) = guard {
-                append!(*ts, "{indent}if ({guard}) {{\n", indent = ctx.indent);
-            }
-            let handler_indent = if guard.is_some() {
-                cstr!("{}  ", ctx.indent)
-            } else {
-                String::from(ctx.indent)
-            };
-
-            let gen_stmt_start = ts.len();
-            if is_implicit_reference {
-                let handler_name = cstr!("__vize_handler_{scope_id}_{}", expr.start);
-                append!(
-                    *ts,
-                    "{indent}const {handler_name} = ((handler: ($event: {event_type}) => unknown) => handler)(({content}));\n",
-                    indent = handler_indent,
-                    event_type = ctx.event_type,
-                );
-                append!(
-                    *ts,
-                    "{indent}{handler_name}($event);  // handler expression\n",
-                    indent = handler_indent,
-                );
-            } else if let Some(event_arg) = inline_callback_event_arg {
-                append!(
-                    *ts,
-                    "{indent}({content})({event_arg});  // handler expression\n",
-                    indent = ctx.indent,
-                );
-            } else {
-                append!(
-                    *ts,
-                    "{indent}{content};  // handler expression\n",
-                    indent = handler_indent
-                );
-            }
-            let gen_stmt_end = ts.len();
-            mappings.push(VizeMapping {
-                gen_range: if is_implicit_reference {
-                    gen_stmt_start..gen_stmt_end
-                } else {
-                    generated_text_range(&ts[gen_stmt_start..gen_stmt_end], content, gen_stmt_start)
-                },
-                src_range: src_start..src_end,
-                sub_spans: Vec::new(),
-            });
-            append!(
-                *ts,
-                "{indent}// @vize-map: handler -> {src_start}:{src_end}\n",
-                indent = handler_indent,
-            );
-            if guard.is_some() {
-                append!(*ts, "{indent}}}\n", indent = ctx.indent);
-            }
-        }
-    }
-}
-
-fn inline_callback_event_argument(content: &str) -> Option<&'static str> {
-    let trimmed = content.trim_start();
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    if let Some(arrow_idx) = trimmed.find("=>") {
-        let before_arrow = strip_async_prefix(trimmed[..arrow_idx].trim_end()).trim();
-        if before_arrow.is_empty() {
-            return None;
-        }
-
-        if let Some(is_empty) = parenthesized_params_are_empty(before_arrow) {
-            return Some(if is_empty { "" } else { "$event" });
-        }
-
-        return is_identifier_segment(before_arrow).then_some("$event");
-    }
-
-    let rest = trimmed.strip_prefix("function")?;
-    let paren_start = trimmed.len() - rest.len() + rest.find('(')?;
-    let paren_end = matching_paren_index(trimmed, paren_start)?;
-    let inner = &trimmed[paren_start + 1..paren_end];
-    Some(if inner.trim().is_empty() {
-        ""
-    } else {
-        "$event"
-    })
-}
-
-fn strip_async_prefix(input: &str) -> &str {
-    let Some(rest) = input.strip_prefix("async") else {
-        return input;
-    };
-    if rest.chars().next().is_some_and(char::is_whitespace) {
-        rest.trim_start()
-    } else {
-        input
-    }
-}
-
-fn parenthesized_params_are_empty(input: &str) -> Option<bool> {
-    if !input.starts_with('(') {
-        return None;
-    }
-    let close = matching_paren_index(input, 0)?;
-    if !input[close + 1..].trim().is_empty() {
-        return None;
-    }
-    Some(input[1..close].trim().is_empty())
-}
-
-fn matching_paren_index(input: &str, open_index: usize) -> Option<usize> {
-    let bytes = input.as_bytes();
-    if bytes.get(open_index) != Some(&b'(') {
-        return None;
-    }
-
-    let mut depth = 0u32;
-    for (idx, byte) in bytes.iter().enumerate().skip(open_index) {
-        match byte {
-            b'(' => depth += 1,
-            b')' => {
-                depth = depth.checked_sub(1)?;
-                if depth == 0 {
-                    return Some(idx);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    None
-}
-
-fn is_callable_handler_reference(content: &str) -> bool {
-    let trimmed = content.trim();
-    if trimmed.is_empty() {
-        return false;
-    }
-
-    let Some(mut idx) = parse_identifier_segment(trimmed, 0) else {
-        return false;
-    };
-
-    loop {
-        idx = skip_ascii_whitespace(trimmed, idx);
-        if idx == trimmed.len() {
-            return true;
-        }
-
-        let rest = &trimmed[idx..];
-        if rest.starts_with("?.[") {
-            idx += 2;
-            let Some(next_idx) = parse_bracket_member(trimmed, idx) else {
-                return false;
-            };
-            idx = next_idx;
-        } else if rest.starts_with("?.") {
-            let Some(next_idx) = parse_identifier_segment(trimmed, idx + 2) else {
-                return false;
-            };
-            idx = next_idx;
-        } else if rest.starts_with('.') {
-            let Some(next_idx) = parse_identifier_segment(trimmed, idx + 1) else {
-                return false;
-            };
-            idx = next_idx;
-        } else if rest.starts_with('[') {
-            let Some(next_idx) = parse_bracket_member(trimmed, idx) else {
-                return false;
-            };
-            idx = next_idx;
-        } else {
-            return false;
-        }
-    }
-}
-
-fn is_identifier_segment(segment: &str) -> bool {
-    let mut chars = segment.chars();
-    let Some(first) = chars.next() else {
-        return false;
-    };
-
-    if !(first == '_' || first == '$' || first.is_alphabetic()) {
-        return false;
-    }
-
-    chars.all(|ch| ch == '_' || ch == '$' || ch.is_alphanumeric())
-}
-
-fn parse_identifier_segment(input: &str, start: usize) -> Option<usize> {
-    let mut chars = input.get(start..)?.char_indices();
-    let (_, first) = chars.next()?;
-    if !is_identifier_start(first) {
-        return None;
-    }
-
-    let mut end = start + first.len_utf8();
-    for (offset, ch) in chars {
-        if !is_identifier_continue(ch) {
-            break;
-        }
-        end = start + offset + ch.len_utf8();
-    }
-    Some(end)
-}
-
-fn is_identifier_start(ch: char) -> bool {
-    ch == '_' || ch == '$' || ch.is_alphabetic()
-}
-
-fn is_identifier_continue(ch: char) -> bool {
-    ch == '_' || ch == '$' || ch.is_alphanumeric()
-}
-
-fn skip_ascii_whitespace(input: &str, mut idx: usize) -> usize {
-    while input
-        .as_bytes()
-        .get(idx)
-        .is_some_and(|byte| byte.is_ascii_whitespace())
-    {
-        idx += 1;
-    }
-    idx
-}
-
-fn parse_bracket_member(input: &str, open_index: usize) -> Option<usize> {
-    if input.as_bytes().get(open_index) != Some(&b'[') {
-        return None;
-    }
-
-    let mut depth = 0u32;
-    let mut quote = None;
-    let mut escaped = false;
-    for (idx, ch) in input
-        .char_indices()
-        .skip_while(|(idx, _)| *idx < open_index)
-    {
-        if let Some(quote_ch) = quote {
-            if escaped {
-                escaped = false;
-            } else if ch == '\\' {
-                escaped = true;
-            } else if ch == quote_ch {
-                quote = None;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' | '"' | '`' => quote = Some(ch),
-            '[' => depth += 1,
-            ']' => {
-                depth = depth.checked_sub(1)?;
-                if depth == 0 {
-                    return Some(idx + ch.len_utf8());
-                }
-            }
-            _ => {}
-        }
-    }
-
-    None
-}
-
-/// Recursively generate child scopes that are VFor/VSlot/EventHandler.
-fn generate_child_scopes(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    ctx: &ScopeGenContext<'_>,
-    parent_scope_id: u32,
-    indent: &str,
-) {
-    if let Some(child_ids) = ctx.children_map.get(&parent_scope_id) {
-        for &child_id in child_ids {
-            if let Some(child_scope) = ctx.summary.scopes.get_scope(child_id)
-                && matches!(
-                    child_scope.kind,
-                    ScopeKind::VFor | ScopeKind::VSlot | ScopeKind::EventHandler
-                )
-            {
-                profile!(
-                    "canon.virtual_ts.scope_node",
-                    generate_scope_node(ts, mappings, ctx, child_scope, indent)
-                );
-            }
-        }
-    }
-}
-
-/// Recursively generate component prop checks inside nested closure scopes (v-for and v-slot).
-fn generate_closure_component_props_recursive(
-    ts: &mut String,
-    mappings: &mut Vec<VizeMapping>,
-    ctx: &VForPropsContext<'_>,
-    scope: &Scope,
-    indent: &str,
-) {
-    let scope_id = scope.id.as_u32();
-    let inner_indent = cstr!("{indent}  ");
-
-    match scope.data() {
-        ScopeData::VFor(data) => {
-            append_v_for_comment(
-                ts,
-                indent,
-                "Component props in v-for scope",
-                data.value_alias.as_str(),
-                data.source.as_str(),
-            );
-            emit_v_for_loop_open(
-                ts,
-                indent,
-                data.value_alias.as_str(),
-                data.key_alias.as_deref(),
-                data.index_alias.as_deref(),
-                data.source.as_str(),
-            );
-
-            // Mark v-for variables as used to avoid TS6133
-            for value in &data.value_bindings {
-                append!(*ts, "{inner_indent}void {value};\n");
-            }
-            if let Some(ref key) = data.key_alias {
-                append!(*ts, "{inner_indent}void {key};\n");
-            }
-            if let Some(ref index) = data.index_alias {
-                append!(*ts, "{inner_indent}void {index};\n");
-            }
-
-            // Emit component prop checks for this scope
-            if let Some(usages) = ctx.components_by_scope.get(&scope_id) {
-                for &(idx, usage) in usages {
-                    profile!(
-                        "canon.virtual_ts.component_prop_checks",
-                        generate_component_prop_checks(
-                            ts,
-                            mappings,
-                            usage,
-                            idx,
-                            ctx.template_prop_names,
-                            ctx.template_offset,
-                            &inner_indent,
-                        )
-                    );
-                }
-            }
-
-            // Recursively handle child closure scopes (v-for and v-slot)
-            if let Some(child_ids) = ctx.children_map.get(&scope_id) {
-                for &child_id in child_ids {
-                    if let Some(child_scope) = ctx.summary.scopes.get_scope(child_id)
-                        && matches!(child_scope.kind, ScopeKind::VFor | ScopeKind::VSlot)
-                    {
-                        profile!(
-                            "canon.virtual_ts.closure_component_props",
-                            generate_closure_component_props_recursive(
-                                ts,
-                                mappings,
-                                ctx,
-                                child_scope,
-                                &inner_indent,
-                            )
-                        );
-                    }
-                }
-            }
-
-            ts.push_str(indent);
-            ts.push_str("});\n");
-        }
-        ScopeData::VSlot(data) => {
-            let props_pattern = data.props_pattern.as_deref().unwrap_or("slotProps");
-            let safe_slot_name = to_safe_identifier_fragment(data.name.as_str());
-            let props_type = slot_props_type(
-                data.component.as_deref(),
-                data.name.as_str(),
-                ctx.summary.scopes.is_v_slot_name_static(scope.id),
-            );
-            append!(
-                *ts,
-                "\n{indent}// Component props in v-slot scope: #{}\n",
-                data.name
-            );
-            append!(
-                *ts,
-                "{indent}void function _slot_props_{safe_slot_name}({props_pattern}: {props_type}) {{\n",
-            );
-            // Mark slot prop variables as used
-            if data.prop_names.is_empty() {
-                append!(*ts, "{inner_indent}void {props_pattern};\n");
-            } else {
-                for prop_name in data.prop_names.iter() {
-                    append!(*ts, "{inner_indent}void {prop_name};\n");
-                }
-            }
-
-            // Emit component prop checks for this scope
-            if let Some(usages) = ctx.components_by_scope.get(&scope_id) {
-                for &(idx, usage) in usages {
-                    profile!(
-                        "canon.virtual_ts.component_prop_checks",
-                        generate_component_prop_checks(
-                            ts,
-                            mappings,
-                            usage,
-                            idx,
-                            ctx.template_prop_names,
-                            ctx.template_offset,
-                            &inner_indent,
-                        )
-                    );
-                }
-            }
-
-            // Recursively handle child closure scopes (v-for and v-slot)
-            if let Some(child_ids) = ctx.children_map.get(&scope_id) {
-                for &child_id in child_ids {
-                    if let Some(child_scope) = ctx.summary.scopes.get_scope(child_id)
-                        && matches!(child_scope.kind, ScopeKind::VFor | ScopeKind::VSlot)
-                    {
-                        profile!(
-                            "canon.virtual_ts.closure_component_props",
-                            generate_closure_component_props_recursive(
-                                ts,
-                                mappings,
-                                ctx,
-                                child_scope,
-                                &inner_indent,
-                            )
-                        );
-                    }
-                }
-            }
-
-            ts.push_str(indent);
-            ts.push_str("};\n");
-        }
-        _ => {}
-    }
-}
